[Refactor] [8/N] to simplify the vLLM openai serving architecture
