+++ vllm/entrypoints/context.py
@@ -24,9 +24,6 @@ from vllm.entrypoints.chat_utils import (
 from vllm.entrypoints.constants import MCP_PREFIX
 from vllm.entrypoints.openai.engine.protocol import (
     FunctionCall,
-    ResponseInputOutputItem,
-    ResponseRawMessageAndToken,
-    ResponsesRequest,
 )
 from vllm.entrypoints.openai.parser.harmony_utils import (
     get_encoding,
@@ -35,6 +32,11 @@ from vllm.entrypoints.openai.parser.harmony_utils import (
 )
 from vllm.entrypoints.openai.parser.responses_parser import (
     get_responses_parser_for_simple_context,
+)
+from vllm.entrypoints.openai.responses.protocol import (
+    ResponseInputOutputItem,
+    ResponseRawMessageAndToken,
+    ResponsesRequest,
 )
 from vllm.entrypoints.responses_utils import construct_tool_dicts
 from vllm.entrypoints.tool import Tool
