+++ vllm/reasoning/abs_reasoning_parsers.py
@@ -19,6 +19,8 @@ if TYPE_CHECKING:
     )
     from vllm.entrypoints.openai.engine.protocol import (
         DeltaMessage,
+    )
+    from vllm.entrypoints.openai.responses.protocol import (
         ResponsesRequest,
     )
     from vllm.tokenizers import TokenizerLike
