+++ vllm/reasoning/basic_parsers.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
     from vllm.entrypoints.openai.chat_completion.protocol import (
         ChatCompletionRequest,
     )
-    from vllm.entrypoints.openai.engine.protocol import (
+    from vllm.entrypoints.openai.responses.protocol import (
         ResponsesRequest,
     )
 else:
