+++ vllm/reasoning/qwen3_reasoning_parser.py
@@ -4,7 +4,9 @@
 from vllm.entrypoints.openai.chat_completion.protocol import (
     ChatCompletionRequest,
 )
-from vllm.entrypoints.openai.engine.protocol import ResponsesRequest
+from vllm.entrypoints.openai.responses.protocol import (
+    ResponsesRequest,
+)
 from vllm.reasoning.basic_parsers import BaseThinkingReasoningParser
